backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
celery==5.4.0
redis==5.1.1
boto3==1.35.30
anthropic==0.34.2
faster-whisper==1.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.10
httpx==0.27.2
stripe==10.11.0
resend==2.4.0
python-slugify==8.0.4
sse-starlette==2.1.3
psycopg2-binary==2.9.9
pypdf==4.3.1
python-docx==1.1.2
